Imports QUnit.only a new feature in QUnit 1.20
