+++ bower.json
@@ -19,7 +19,7 @@
     "ember-test-helpers": "^0.5.14"
   },
   "devDependencies": {
-    "qunit": "^1.17.1",
+    "qunit": "^1.20.0",
     "loader": "stefanpenner/loader.js#1.0.1",
     "ember-cli-shims": "stefanpenner/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "rwjblue/ember-cli-test-loader#0.0.4",
