+++ lib/ember-qunit.js
@@ -2,6 +2,7 @@ import moduleFor          from 'ember-qunit/module-for';
 import moduleForComponent from 'ember-qunit/module-for-component';
 import moduleForModel     from 'ember-qunit/module-for-model';
 import test               from 'ember-qunit/test';
+import only               from 'ember-qunit/only';
 import { setResolver }    from 'ember-test-helpers';
 
 export {
@@ -9,5 +10,6 @@ export {
   moduleForComponent,
   moduleForModel,
   test,
+  only,
   setResolver
 };
