+++ lib/ember-qunit/only.js
@@ -1,0 +1,6 @@
+import testWrapper from 'ember-qunit/test-wrapper';
+import { only as qunitOnly } from 'qunit';
+
+export default function only(testName, callback) {
+  testWrapper(testName, callback, qunitOnly);
+}
