+++ lib/qunit.js
@@ -3,5 +3,6 @@
 export var module = QUnit.module;
 export var test = QUnit.test;
 export var skip = QUnit.skip;
+export var only = QUnit.only;
 
 export default QUnit;
