+++ tests/qunit-shim-test.js
@@ -1,7 +1,8 @@
 import {
   module,
   test,
-  skip
+  skip,
+  only
 } from 'qunit';
 
 module('qunit-shim test');
@@ -12,4 +13,8 @@ test('should work!', function(assert) {
 
 test('imports skips', function(assert) {
   assert.equal(typeof skip, 'function', 'imports QUnit.skip');
+});
+
+test('imports only', function(assert) {
+  assert.equal(typeof only, 'function', 'imports QUnit.only');
 });
